Everything: Code cleanup
- Removed deprecated versions of Classifier Trainer and Processor boxes
- Removed remaining references to unused BliFF
- Removed Mean Variance box: undocumented and with side effects
- Cleaned up some old packages from linux dependency installer
- Fixed a compiler nag


Former-commit-id: 366d6360ba30393b2042665921c8482998dcff36

+++ applications/platform/acquisition-server/src/drivers/labstreaminglayer/ovasCConfigurationLabStreamingLayer.cpp
@@ -67,7 +67,7 @@ boolean CConfigurationLabStreamingLayer::preConfigure(void)
 	boolean l_bExactSignalMatch = false; 	// If we can match both name and ID, stop at that. Otherwise we auto-accept the 'name only' match.
 	boolean l_bExactMarkerMatch = false;
 
-	m_rDriverContext.getLogManager() << LogLevel_Trace << "Discovered " << m_vStreams.size() << " streams in total\n";
+	m_rDriverContext.getLogManager() << LogLevel_Trace << "Discovered " << static_cast<uint64>(m_vStreams.size()) << " streams in total\n";
 
 	for(uint32 i=0; i<m_vStreams.size(); i++)
 	{
